+++ src/algorithms/utils.tpp
@@ -5,25 +5,25 @@
 
 using std::vector;
 
-bool is_power2(long x) {
+bool is_power2(long long x) {
 	return (x != 0) && ((x & (x - 1)) == 0);
 }
 
-long flatten(long i , long j) {
+long long flatten(long long i , long long j) {
 	return (1<<i)+j;
 }
 
 template <typename T>
-vector<T> get_subtree (const vector<T>& tree, long k) {
-	long N = tree.size();
+vector<T> get_subtree (const vector<T>& tree, long long k) {
+	long long N = tree.size();
 
 	vector<T> subtree(N);
 
-	long i = 0;
+	long long i = 0;
 	while ((1<<i)*(k+1) <= N) { //up through the leaves 
-		for (long j = 0; j < (1<<i); ++j) {
+		for (long long j = 0; j < (1<<i); ++j) {
 
-			long node = flatten(i,j);
+			long long node = flatten(i,j);
 			subtree[node] = tree[node];
 		}
 		++i;
@@ -34,8 +34,8 @@ vector<T> get_subtree (const vector<T>& tree, long k) {
 }
 
 template <typename T>
-vector<T> get_shade(const vector<T>& vals, long k) {
-	long N = vals.size();
+vector<T> get_shade(const vector<T>& vals, long long k) {
+	long long N = vals.size();
 
 	assert (k < 2*N);
 
@@ -45,8 +45,8 @@ vector<T> get_shade(const vector<T>& vals, long k) {
 		return slice;
 	}
 
-	long left = k; //this will be the leftmost leaf under the subtree rooted at P
-	long right = k; //similarly the rightmost leaf under subtree rooted at P
+	long long left = k; //this will be the leftmost leaf under the subtree rooted at P
+	long long right = k; //similarly the rightmost leaf under subtree rooted at P
 	while (left < N) { //stops when left_span is a leaf
 		left = 2*left;
 		right = 2*right + 1;
@@ -65,11 +65,11 @@ vector<T> get_shade(const vector<T>& vals, long k) {
 
 template<typename T>
 vector<T> elt_to_base(const vector<Elt<T>> &e) {
-	long N = e.size();
+	long long N = e.size();
 
 	vector<T> b(N);
 
-	for(long i = 0; i < N; ++i) {
+	for(long long i = 0; i < N; ++i) {
 		b[i] = e[i].t;
 	}
 
@@ -78,11 +78,11 @@ vector<T> elt_to_base(const vector<Elt<T>> &e) {
 
 template<typename T>
 vector<Elt<T>> base_to_elt(const vector<T>& b) {
-	long N = b.size();
+	long long N = b.size();
 
 	vector<Elt<T>> e(N);
 
-	for(long i = 0; i < N; ++i) {
+	for(long long i = 0; i < N; ++i) {
 		e[i] = b[i];
 	}
