+++ src/searches/benchmarks.hpp
@@ -1,13 +1,13 @@
 #ifndef REMAINDERTREE_SRC_SEARCHES_BENCHMARKS_H_
 #define REMAINDERTREE_SRC_SEARCHES_BENCHMARKS_H_
 
-//std::function< std::vector<bool> (long N)> remember this is what is being passed in!
+//std::function< std::vector<bool> (long long N)> remember this is what is being passed in!
 #include <vector>
 
 
-std::vector<bool> constant_slow(long);
+std::vector<bool> constant_slow(long long);
 
-std::vector<bool> random_zz(long, long, long);
+std::vector<bool> random_zz(long long, long long, long long);
 
 
 #endif //REMAINDERTREE_SRC_SEARCHES_BENCHMARKS_H_
